Move keyboard code into touch number input
