+++ UIElements/touchNumberInput.py
@@ -6,6 +6,7 @@ This allows the user to touch or keyboard input a number when it is the content 
 from   kivy.uix.gridlayout                       import   GridLayout
 from   kivy.properties                           import   ObjectProperty
 from   kivy.properties                           import   StringProperty
+import global_variables
 
 class TouchNumberInput(GridLayout):
     done   = ObjectProperty(None)
@@ -15,6 +16,43 @@ class TouchNumberInput(GridLayout):
         super(TouchNumberInput,self).__init__(**kwargs)
         
         self.unitsBtn.text = self.data.units
+        
+        if global_variables._keyboard:
+            global_variables._keyboard.bind(on_key_down=self.keydown_popup)
+            self.bind(on_dismiss=self.ondismiss_popup)
+        
+    
+    def keydown_popup(self, keyboard, keycode, text, modifiers):
+        if (keycode[1] == '0') or (keycode[1] =='numpad0'):
+            self.addText('0')
+        elif (keycode[1] == '1') or (keycode[1] =='numpad1'):
+            self.addText('1')
+        elif (keycode[1] == '2') or (keycode[1] =='numpad2'):
+            self.addText('2')
+        elif (keycode[1] == '3') or (keycode[1] =='numpad3'):
+            self.addText('3')
+        elif (keycode[1] == '4') or (keycode[1] =='numpad4'):
+            self.addText('4')
+        elif (keycode[1] == '5') or (keycode[1] =='numpad5'):
+            self.addText('5')
+        elif (keycode[1] == '6') or (keycode[1] =='numpad6'):
+            self.addText('6')
+        elif (keycode[1] == '7') or (keycode[1] =='numpad7'):
+            self.addText('7')
+        elif (keycode[1] == '8') or (keycode[1] =='numpad8'):
+            self.addText('8')
+        elif (keycode[1] == '9') or (keycode[1] =='numpad9'):
+            self.addText('9')
+        elif (keycode[1] == '.') or (keycode[1] =='numpaddecimal'):
+            self.addText('.')
+        elif (keycode[1] == 'backspace'):
+            self.textInput.text = self.textInput.text[:-1]         
+        elif (keycode[1] == 'enter') or (keycode[1] =='numpadenter'):
+            self.done()
+        elif (keycode[1] == 'escape'):     # abort entering a number, keep the old number
+            self.textInput.text = ''    # clear text so it isn't converted to a number
+            self.done()
+        return True     # always swallow keypresses since this is a modal dialog
     
     def addText(self, text):
         '''
@@ -36,3 +74,7 @@ class TouchNumberInput(GridLayout):
         else:
             self.data.units = "INCHES"
             self.unitsBtn.text = 'INCHES'
+    
+    def ondismiss_popup(self, event):
+        if global_variables._keyboard:
+            global_variables._keyboard.unbind(on_key_down=self.keydown_popup)
